requirements.txt
# GAELP Training Orchestrator Requirements
# Core dependencies for the training orchestrator system

# Async and concurrency
asyncio-throttle==1.0.2
aioredis==2.0.1

# Google Cloud Platform
google-cloud-bigquery==3.11.4
google-cloud-pubsub==2.18.1
google-cloud-storage==2.10.0

# Data processing and ML
numpy==1.24.3
scipy==1.11.1
pandas==2.0.3
scikit-learn==1.3.0
statsmodels>=0.14.0
faker>=19.0.0
joblib>=1.3.0

# Offline RL
d3rlpy==2.8.1
torch>=2.5.0
gymnasium==1.0.0
h5py
matplotlib
seaborn

# Redis for state management
redis==4.6.0
hiredis==2.2.3

# Configuration and serialization
pydantic==2.1.1
PyYAML==6.0.1
toml==0.10.2

# Monitoring and logging
prometheus-client==0.17.1
structlog==23.1.0

# gRPC for high-throughput communication
grpcio==1.56.2
grpcio-tools==1.56.2
protobuf==4.23.4

# HTTP clients and utilities
httpx==0.24.1
requests==2.31.0

# Testing (for development)
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-mock==3.11.1
pytest-cov==4.1.0

# Development tools
black==23.7.0
isort==5.12.0
flake8==6.0.0
mypy==1.5.1

# Documentation (optional)
sphinx==7.1.2
sphinx-rtd-theme==1.3.0

# Environment management
python-dotenv==1.0.0

# Type hints
typing-extensions==4.7.1

# RecSim for user behavior simulation
recsim-ng==0.2.0
tensorflow==2.13.0
dm-tree==0.1.8

# Utilities
click==8.1.6
rich==13.5.2
tabulate==0.9.0

# RL and Gymnasium Environment
gymnasium>=1.2.0
stable-baselines3>=2.7.0
matplotlib>=3.7.0

# MCP and Ad Platform Integration
mcp==1.13.0
modelcontextprotocol==0.1.0
meta-ads-mcp==0.11.2
tiktok-ads-mcp==0.1.2
google-ads==28.0.0
google-auth-oauthlib==1.2.2

# Experiment tracking
wandb==0.21.1
psutil==7.0.0

# Model Versioning Dependencies
GitPython==3.1.40
pathlib
dataclasses
hashlib2

# Delayed Reward System Dependencies
sqlalchemy==2.0.20
alembic==1.12.0

# Production Ad Account Management - REAL MONEY SYSTEMS
facebook-business==19.0.0
google-ads==26.0.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-analytics-data==0.18.11

# Budget monitoring and alerts
twilio==8.10.0
schedule==1.2.0

# Security and encryption for credential storage
cryptography==41.0.7
keyring==24.2.0